ending refactoring Calendar component
