+++ public/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link href='https://fonts.googleapis.com/css?family=Open Sans' rel='stylesheet' />
-    <title>AsideTask Manager</title>
+    <title>Task Manager</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
