+++ src/components/AsideBar/SupportComponents/AsideTask/AsideTask.module.scss
@@ -1,3 +1,5 @@
+@import "/src/variable";
+
 .Task {
   display: flex;
   flex-direction: column;
