+++ src/components/Calendar/SupportComponents/CalendarCell/CalendarCell.module.scss
@@ -1,0 +1,10 @@
+.CalendarCell__title{
+  font-size: 1.5rem;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+
+  padding-right: 0.1rem;
+}
